cryptfs_hw: Remove android reboot during password check
After reaching the maximum number of retires with wrong password,
userdata is erased based on the info written on /cache partition.
As per the latest changes of A/B, cache partition is not present
which leads to failure of userdata wipe before reboot. This patch
removes reboot from cryptfs_hw layer if max entries reached, as
frameworks take care of userdata wipe.

Change-Id: If97f2f7452beb004f44d7d41d1d76def23fe2720